This PR supercedes #2039, which had the following description:

> Support for parallel compilation of *.erl file was dropped before 3.0 release.
> However, our tests for a project containing ~500 source files show substantial gain, lowering compilation time from 58 seconds to 18 on a MacBook Pro 15" (4 cores, 8 threads), and to just 10 seconds on Xeon-D machine.

This patch does two things on top of @max-au's PR:

 1. it broadens the interface for the compiler module so that non-first-file modules can possibly be parallelized. This is done by dynamically switching on `[ListOfFiles]`, which remains sequential as
       before, or `{[SeqPriority], [Parallel]}`, which divides regular files between higher priority ones and those that can be in parallel
2. implements this mechanism in the rebar compiler, based on the erl file digraph. If a file has an in-neighbour, it is depended on by another file. The mechanism therefore makes it so all files that have       dependants get compiled in their strict relative sequential order first, and then the undepended-on files get compiled together in parallel.

By running:

      ./rebar3 ct --suite test/rebar_compile_SUITE.erl --case recompile_when_parse_transform_inline_changes --repeat 50

the previous iteration of this PR would rapidly fail, and this one succeeds every time.