@ferd 2019-03-29
Running this branch on Rebar3 itself had low impact on compile performance:

- this branch: `rebar3 escriptize  10.73s user 2.45s system 181% cpu 7.275 total`
- master: `rebar3 escriptize  10.48s user 2.53s system 116% cpu 11.171 total`

The slight slowdown on a small-enough project might certainly be worth the gain on a large one.