@max-au 2019-03-29
I like this one. Performance results: 
time rebar3 compile real	0m14.887s
Before the patch: 
time rebar3 compile real	0m45.576s
Thank you for fixing!