@ferd 2019-03-29
@tsloughter re: your comment on the other thread. I think things are not _too_ bad considering there's a check for `min(length(Files), NumScheduler)`. The old version always defaulted to 3 workers no matter the scheduler count nor file count: https://github.com/erlang/rebar3/pull/265/files#diff-2015be9e4c4af12364f637580b5450bdL52

I had limited slow downs here as well compared to the massive speedup Max has reported, but we might want to check with smaller projects to find a cutoff point.