@tsloughter 2019-03-29
Maybe it should not spawn any workers if the `min` is `< NumSchedulers`?