@max-au 2019-03-29
Now even cheap desktops have 16+ threads, and parallel build with 8 files (< NumSchedulers) benefits from using 8 workers. In fact, spawning a worker does not not cost much, spawn rate of 8,000 per second is fine on a laptop.