@max-au 2019-03-30
After testing a bit more, I realised that there is also a bottleneck in init_dag, which is single-threaded, does some file i/o, and is quite slow as it performs full-scale *.erl file parsing (rebar_compiler_erl, line 242).
That one would be a bit harder to parallelise, but it's still worth trying.

==
(unrelated - diagnostic message produced when myapp.app file is broken - line 346 of rebar_app_discover, create_app_info - {badmatch, ...} - quite misleading).